pyproject.toml
[project]
name = "proj-riot"
version = "0.1.0"
description = "Projeto gerenciado com uv e pyproject.toml"
authors = [
    {name = "Helton Maia", email = "helton.maia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib==3.10.0",
    "numpy==1.26.4",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-cusparselt-cu12==0.6.2",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "open-clip-torch==2.32.0",
    "opencv-python==4.11.0.86",
    "pandas==2.2.3",
    "pillow==10.4.0",
    "pyyaml==6.0.2",
    "scikit-learn==1.6.1",
    "scipy==1.15.1",
    "seaborn==0.13.2",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "torch==2.6.0",
    "torchvision==0.21.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
